refactor(db): relocate database folder
