+++ src/models/catalog.model.js
@@ -1,5 +1,5 @@
 import { DataTypes } from 'sequelize';
-import sequelize from '../../db/database.js';
+import sequelize from '../db/database.js';
 
 const Catalog = sequelize.define('Catalog', {
   name: {
