+++ src/models/control.model.js
@@ -1,5 +1,5 @@
 import { DataTypes } from 'sequelize';
-import sequelize from '../../db/database.js';
+import sequelize from '../db/database.js';
 
 const Control = sequelize.define(
   'Control',
