+++ src/models/models.js
@@ -1,7 +1,7 @@
 import { Sequelize } from 'sequelize';
-import sequelize from '../../db/database.js';
-import applyExtraSetup from '../../db/extra-setup.js';
-import { associateModels } from '../../db/model-utils.js';
+import sequelize from '../db/database.js';
+import applyExtraSetup from '../db/extra-setup.js';
+import { associateModels } from '../db/model-utils.js';
 import User from './user.model.js';
 import Assistant from './assistant.model.js';
 import Thread from './thread.model.js';
